edit all

+++ db/migrate/20210705111125_create_reviews.rb
@@ -3,7 +3,7 @@ class CreateReviews < ActiveRecord::Migration[6.1]
     create_table :reviews do |t|
       t.belongs_to :user, null: false, foreign_key: true
       t.belongs_to :product, null: false, foreign_key: true
-      t.float :rate
+      t.integer :rate
       t.text :text
 
       t.timestamps
